Bump gem versions for newer rubies
In order for some tests to pass on Ruby 2.4 or JRuby 9.2.0.0 and higher, Arel must be at least version 6.0.4 to account for breaking changes to Fixnum/Bignum/Integer.  This also means the version of ActiveRecord must be raised; all versions before 4.2 require older versions of Arel.  In effect, this raises the minimum requirement to Rails to 4.2 for users wishing to use the Rails integration on newer Rubies.